requirements.txt
customtkinter==5.2.1
darkdetect==0.8.0
packaging==23.2
pillow==10.2.0
